python/requirements.txt
redis==5.0.1
pybind11==3.0.2
websocket-client==1.8.0